refactor(models): reorganize into logical sub-packages with unified exports.
- Split models package into domain-specific sub-packages:

1. core: Exercise, ExerciseType, MuscleGroup, SetType, and more will be added later.
2. planned: PlannedSet, and more will be added later.
3. performed: stub for now, and more will be added later.
4. user: User, and more will be added later.

- Move Base model class from package __init__ to base.py.
- Add explicit exports in subpackage __init__ files.
- Update test imports to reflect new structure.
