+++ styx/models/core/__init__.py
@@ -1,0 +1,6 @@
+from .exercise import Exercise
+from .exercise_type import ExerciseType
+from .muscle_group import MuscleGroup
+from .set_type import SetType
+
+__all__ = ["Exercise", "ExerciseType", "MuscleGroup", "SetType"]
