+++ styx/models/core/exercise_type.py
@@ -1,9 +1,13 @@
-from sqlalchemy import List, Optional, String
+from typing import TYPE_CHECKING
+
+from sqlalchemy import String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from . import BaseModel
-from .exercise_model import Exercise
-from .relationships import exercise_type_association_table
+from ..base import BaseModel
+from ..relationships import exercise_type_association_table
+
+if TYPE_CHECKING:
+    from .exercise import Exercise
 
 
 class ExerciseType(BaseModel):
@@ -12,7 +16,7 @@ class ExerciseType(BaseModel):
     name: Mapped[str] = mapped_column(String(120), unique=True)
 
     # many to many field with Exercise table
-    exercises: Mapped[Optional[List[Exercise]]] = relationship(
+    exercises: Mapped[list["Exercise"] | None] = relationship(
         "Exercise",
         secondary=exercise_type_association_table,
         back_populates="exercise_types",
@@ -22,4 +26,3 @@ class ExerciseType(BaseModel):
 
     def __repr__(self) -> str:
         return f"ExerciseType(name={self.name})"
-
