+++ styx/models/planned/__init__.py
@@ -1,0 +1,3 @@
+from .planned_set import PlannedSet
+
+__all__ = ["PlannedSet"]
