+++ styx/models/user/__init__.py
@@ -1,0 +1,3 @@
+from .user import User
+
+__all__ = ["User"]
